Tidy up types
