+++ src/check-stack-children.ts
@@ -1,54 +1,6 @@
-import { Rule } from "eslint";
-import { BaseNode } from "estree";
+import { RuleModule } from "./eslint-types";
 
-/**
- * This is all required because the JSXElement is included in eslint by default
- * (via acorn-jsx) but is not in the typescript definitions.
- */
-
-type JSXElementChildNode = JSXElementNode | JSXExpressionContainerNode;
-
-interface JSXExpressionContainerNode {
-  type: "JSXExpressionContainer";
-  expression: JSXElementNode;
-}
-
-interface JSXOpeningElementNode {
-  type: "JSXOpeningElement";
-  name: JSXIdentifierNode;
-  selfClosing: boolean;
-}
-
-interface JSXIdentifierNode {
-  type: "JSXIdentifier";
-  name: string;
-}
-interface JSXElementNode extends BaseNode {
-  id: string;
-  type: "JSXElement";
-  openingElement: JSXOpeningElementNode;
-  children: JSXElementChildNode[];
-}
-
-declare module "eslint" {
-  namespace Rule {
-    interface NodeListener {
-      JSXElement(node: JSXElementNode): void;
-    }
-
-    interface RuleModule {
-      create(context: RuleContext): NodeListener;
-    }
-  }
-}
-
-declare module "estree" {
-  interface NodeMap {
-    JSXElement: JSXElementNode;
-  }
-}
-
-export const rules: Record<"no-redundant-stack", Rule.RuleModule> = {
+export const rules: Record<"no-redundant-stack", RuleModule> = {
   "no-redundant-stack": {
     meta: {
       docs: {
@@ -62,8 +14,8 @@ export const rules: Record<"no-redundant-stack", Rule.RuleModule> = {
       },
     },
     create: (context) => {
-      const listener = {
-        JSXElement: (node: JSXElementNode) => {
+      return {
+        JSXElement: (node) => {
           const elementName = node.openingElement.name.name;
           if (elementName !== "Stack") {
             return;
@@ -98,8 +50,7 @@ export const rules: Record<"no-redundant-stack", Rule.RuleModule> = {
             });
           }
         },
-      } satisfies Rule.NodeListener;
-      return listener;
+      };
     },
   },
 };
